+++ lib/appfuel/initializers/all.rb
@@ -1,3 +1,3 @@
-require_relative 'logger'
+require_relative 'logging'
 require_relative 'db'
 require_relative 'web_api'